% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedsuite-package.R
\docType{package}
\name{pedsuite-package}
\alias{pedsuite}
\alias{pedsuite-package}
\title{pedsuite: Easy Installation of the 'ped suite' Packages for Pedigree Analysis}
\description{
The 'ped suite' is a collection of packages for pedigree analysis, covering applications in forensic genetics, medical genetics and more. A detailed presentation of the 'ped suite' is given in the book 'Pedigree Analysis in R' (Vigeland, 2021, ISBN: 9780128244302).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://magnusdv.github.io/pedsuite/}
  \item \url{https://github.com/magnusdv/pedsuite}
  \item Report bugs at \url{https://github.com/magnusdv/pedsuite/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

Authors:
\itemize{
  \item Thore Egeland \email{thore.egeland@nmbu.no} (\href{https://orcid.org/0000-0002-3465-8885}{ORCID})
}

}
\keyword{internal}
