% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedsuite-package.R
\docType{package}
\name{pedsuite}
\alias{pedsuite}
\title{pedsuite: A collection of packages for pedigree analysis in R}
\description{
This package makes it easy to install and load the \strong{ped suite} packages for
pedigree analysis in R, covering applications in forensic genetics, medical
genetics and others. A detailed presentation of the \strong{ped suite} is given in
the book \emph{Pedigree Analysis in R} (Vigeland, 2021, ISBN: 9780128244302).
}
