% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedsuite-package.R
\docType{package}
\name{pedsuite}
\alias{pedsuite}
\title{pedsuite: A collection of packages for pedigree analysis in R}
\description{
The 'ped suite' is a collection of packages for pedigree analysis, covering
applications in forensic genetics, medical genetics and others. A detailed
presentation of the 'ped suite' is given in the book 'Pedigree Analysis in R'
(Vigeland, 2021, ISBN: 9780128244302).
}
